requirements.txt
flask==3.1.3
flask-cors==6.0.2
svglib==1.6.0
reportlab==4.4.10
Pillow==12.1.1
# cadquery installed via conda: conda install -c cadquery -c conda-forge cadquery
